+++ src/main/java/com/telepathicgrunt/worldblender/configs/WBBlendingConfigs.java
@@ -59,7 +59,7 @@ public class WBBlendingConfigs implements ConfigData {
              and vanilla's icebergs:
              "tree, minecraft:iceberg"
             """)
-    public String blacklistedFeatures = "minecraft:basalt_blobs, minecraft:blackstone_blobs";
+    public String blacklistedFeatures = "minecraft:basalt_blobs, minecraft:blackstone_blobs, betterend:sulphuric_lake, betterend:end_lake_rare, betterend:end_lake_normal, betterend:end_lake, betterend:desert_lake, betterend:tunel_cave, betterend:empty_aurora_cave_cave_populator, betterend:empty_end_cave_cave_populator, betterend:empty_smaragdant_cave_cave_populator, betterend:jade_cave_cave_populator, betterend:lush_aurora_cave_cave_populator, betterend:lush_smaragdant_cave_cave_populator";
 
 
     @ConfigEntry.Gui.PrefixText
@@ -82,7 +82,7 @@ public class WBBlendingConfigs implements ConfigData {
              and vanilla's igloos:
              "temple, minecraft:igloo"
             """)
-    public String blacklistedStructures = "";
+    public String blacklistedStructures = "betterend:painted_mountain, betterend:mountain, betterend:megalake, betterend:megalake_small";
 
 
     @ConfigEntry.Gui.PrefixText